requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
blinker==1.9.0
blis==1.2.0
boto3==1.36.26
botocore==1.36.26
browser-cookie3==0.20.1
bs4==0.0.2
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cloudpathlib==0.20.0
confection==0.1.5
cymem==2.0.11
decorator==5.2.0
deprecation==2.1.0
distro==1.9.0
easyocr==1.7.2
fastapi==0.115.8
ffmpeg-python==0.2.0
filelock==3.17.0
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
gotrue==2.11.4
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
llvmlite==0.44.0
lz4==4.4.3
marisa-trie==1.2.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
multidict==6.1.0
murmurhash==1.0.12
narwhals==1.27.1
networkx==3.4.2
ninja==1.11.1.3
numba==0.61.0
numpy==2.1.3
openai==1.64.0
openai-whisper==20240930
opencv-python-headless==4.11.0.86
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
pip==25.0
playwright==1.50.0
postgrest==0.19.3
preshed==3.0.9
proglog==0.1.10
propcache==0.3.0
protobuf==5.29.3
pyarrow==19.0.1
pyclipper==1.3.0.post6
pycryptodomex==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
pyktok==0.0.31
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
realtime==2.4.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.23.1
s3transfer==0.11.2
scikit-image==0.25.2
scipy==1.15.2
selenium==4.29.0
setuptools==75.8.0
shapely==2.0.7
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.5.1
starlette==0.45.3
storage3==0.11.3
streamlit==1.42.2
StrEnum==0.4.15
supabase==2.13.0
supafunc==0.9.3
sympy==1.13.1
tenacity==9.0.0
thinc==8.3.4
tifffile==2025.2.18
TikTokApi==7.0.0
tiktoken==0.9.0
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
wasabi==1.1.3
weasel==0.4.1
websocket-client==1.8.0
websockets==14.2
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
yt-dlp==2025.2.19
